+++ pgrx-tests/src/tests/aggregate_tests.rs
@@ -10,7 +10,6 @@
 use pgrx::prelude::*;
 use pgrx::Internal;
 use serde::{Deserialize, Serialize};
-use std::collections::HashSet;
 
 #[derive(Copy, Clone, Default, Debug, PostgresType, Serialize, Deserialize)]
 pub struct DemoSum {
@@ -61,49 +60,6 @@ impl Aggregate for DemoSum {
     ) -> Self::State {
         first += second;
         first
-    }
-}
-
-#[derive(Copy, Clone, Default, Debug)]
-pub struct DemoUnique;
-
-#[pg_aggregate]
-impl Aggregate for DemoUnique {
-    type Args = &'static str;
-    type State = Internal;
-    type Finalize = i32;
-
-    fn state(
-        mut current: Self::State,
-        arg: Self::Args,
-        _fcinfo: pg_sys::FunctionCallInfo,
-    ) -> Self::State {
-        let inner = unsafe { current.get_or_insert_default::<HashSet<String>>() };
-
-        inner.insert(arg.to_string());
-        current
-    }
-
-    fn combine(
-        mut first: Self::State,
-        mut second: Self::State,
-        _fcinfo: pg_sys::FunctionCallInfo,
-    ) -> Self::State {
-        let first_inner = unsafe { first.get_or_insert_default::<HashSet<String>>() };
-        let second_inner = unsafe { second.get_or_insert_default::<HashSet<String>>() };
-
-        let unioned: HashSet<_> = first_inner.union(second_inner).collect();
-        Internal::new(unioned)
-    }
-
-    fn finalize(
-        mut current: Self::State,
-        _direct_args: Self::OrderedSetArgs,
-        _fcinfo: pg_sys::FunctionCallInfo,
-    ) -> Self::Finalize {
-        let inner = unsafe { current.get_or_insert_default::<HashSet<String>>() };
-
-        inner.len() as i32
     }
 }
