+++ pgrx-tests/src/tests/heap_tuple.rs
@@ -12,7 +12,7 @@ const DOG_COMPOSITE_TYPE: &str = "Dog";
 
 use pgrx::pgbox::AllocatedByRust;
 use pgrx::prelude::*;
-use pgrx::{Aggregate, VariadicArray};
+use pgrx::VariadicArray;
 
 extension_sql!(
     r#"
@@ -555,76 +555,6 @@ mod sql_generator_tests {
         ),
     > {
         TableIterator::once((None, None))
-    }
-
-    #[derive(Copy, Clone, Default, Debug)]
-    pub struct AggregateWithOrderedSetArgs;
-
-    #[pg_aggregate]
-    impl Aggregate for AggregateWithOrderedSetArgs {
-        type Args = name!(input, pgrx::composite_type!('static, "Dog"));
-        type State = pgrx::composite_type!('static, "Dog");
-        type Finalize = pgrx::composite_type!('static, "Dog");
-        const ORDERED_SET: bool = true;
-        type OrderedSetArgs = name!(percentile, pgrx::composite_type!('static, "Dog"));
-
-        fn state(
-            _current: Self::State,
-            _arg: Self::Args,
-            _fcinfo: pg_sys::FunctionCallInfo,
-        ) -> Self::State {
-            unimplemented!("Just a SQL generation test")
-        }
-
-        fn finalize(
-            _current: Self::State,
-            _direct_arg: Self::OrderedSetArgs,
-            _fcinfo: pg_sys::FunctionCallInfo,
-        ) -> Self::Finalize {
-            unimplemented!("Just a SQL generation test")
-        }
-    }
-
-    #[derive(Copy, Clone, Default, Debug)]
-    pub struct AggregateWithMovingState;
-
-    #[pg_aggregate]
-    impl Aggregate for AggregateWithMovingState {
-        type Args = pgrx::composite_type!('static, "Dog");
-        type State = pgrx::composite_type!('static, "Dog");
-        type MovingState = pgrx::composite_type!('static, "Dog");
-
-        fn state(
-            _current: Self::State,
-            _arg: Self::Args,
-            _fcinfo: pg_sys::FunctionCallInfo,
-        ) -> Self::State {
-            unimplemented!("Just a SQL generation test")
-        }
-
-        fn moving_state(
-            _current: Self::State,
-            _arg: Self::Args,
-            _fcinfo: pg_sys::FunctionCallInfo,
-        ) -> Self::MovingState {
-            unimplemented!("Just a SQL generation test")
-        }
-
-        fn moving_state_inverse(
-            _current: Self::State,
-            _arg: Self::Args,
-            _fcinfo: pg_sys::FunctionCallInfo,
-        ) -> Self::MovingState {
-            unimplemented!("Just a SQL generation test")
-        }
-
-        fn combine(
-            _first: Self::State,
-            _second: Self::State,
-            _fcinfo: pg_sys::FunctionCallInfo,
-        ) -> Self::State {
-            unimplemented!("Just a SQL generation test")
-        }
     }
 
     #[pg_extern]
