+++ pgrx-tests/tests/compile-fail/aggregate-functions-dont-run-forever.rs
@@ -1,5 +1,7 @@
 use pgrx::aggregate::*;
 use pgrx::prelude::*;
+use pgrx::Internal;
+use std::collections::HashSet;
 
 const DOG_COMPOSITE_TYPE: &str = "Dog";
 
@@ -59,6 +61,119 @@ impl Aggregate for ScritchCollector {
         let current_scritches: i32 = current.get_by_name("scritches").unwrap().unwrap_or_default();
         current.set_by_name("scritches", current_scritches + arg).unwrap();
         Some(current)
+    }
+}
+
+#[derive(Copy, Clone, Default, Debug)]
+pub struct DemoUnique;
+
+#[pg_aggregate]
+impl Aggregate for DemoUnique {
+    type Args = &'static str;
+    type State = Internal;
+    type Finalize = i32;
+
+    fn state(
+        mut current: Self::State,
+        arg: Self::Args,
+        _fcinfo: pg_sys::FunctionCallInfo,
+    ) -> Self::State {
+        let inner = unsafe { current.get_or_insert_default::<HashSet<String>>() };
+
+        inner.insert(arg.to_string());
+        current
+    }
+
+    fn combine(
+        mut first: Self::State,
+        mut second: Self::State,
+        _fcinfo: pg_sys::FunctionCallInfo,
+    ) -> Self::State {
+        let first_inner = unsafe { first.get_or_insert_default::<HashSet<String>>() };
+        let second_inner = unsafe { second.get_or_insert_default::<HashSet<String>>() };
+
+        let unioned: HashSet<_> = first_inner.union(second_inner).collect();
+        Internal::new(unioned)
+    }
+
+    fn finalize(
+        mut current: Self::State,
+        _direct_args: Self::OrderedSetArgs,
+        _fcinfo: pg_sys::FunctionCallInfo,
+    ) -> Self::Finalize {
+        let inner = unsafe { current.get_or_insert_default::<HashSet<String>>() };
+
+        inner.len() as i32
+    }
+}
+
+#[derive(Copy, Clone, Default, Debug)]
+pub struct AggregateWithOrderedSetArgs;
+
+#[pg_aggregate]
+impl Aggregate for AggregateWithOrderedSetArgs {
+    type Args = name!(input, pgrx::composite_type!('static, "Dog"));
+    type State = pgrx::composite_type!('static, "Dog");
+    type Finalize = pgrx::composite_type!('static, "Dog");
+    const ORDERED_SET: bool = true;
+    type OrderedSetArgs = name!(percentile, pgrx::composite_type!('static, "Dog"));
+
+    fn state(
+        _current: Self::State,
+        _arg: Self::Args,
+        _fcinfo: pg_sys::FunctionCallInfo,
+    ) -> Self::State {
+        unimplemented!("Just a SQL generation test")
+    }
+
+    fn finalize(
+        _current: Self::State,
+        _direct_arg: Self::OrderedSetArgs,
+        _fcinfo: pg_sys::FunctionCallInfo,
+    ) -> Self::Finalize {
+        unimplemented!("Just a SQL generation test")
+    }
+}
+
+#[derive(Copy, Clone, Default, Debug)]
+pub struct AggregateWithMovingState;
+
+#[pg_aggregate]
+impl Aggregate for AggregateWithMovingState {
+    type Args = pgrx::composite_type!('static, "Dog");
+    type State = pgrx::composite_type!('static, "Dog");
+    type MovingState = pgrx::composite_type!('static, "Dog");
+
+    fn state(
+        _current: Self::State,
+        _arg: Self::Args,
+        _fcinfo: pg_sys::FunctionCallInfo,
+    ) -> Self::State {
+        unimplemented!("Just a SQL generation test")
+    }
+
+    fn moving_state(
+        _current: Self::State,
+        _arg: Self::Args,
+        _fcinfo: pg_sys::FunctionCallInfo,
+    ) -> Self::MovingState {
+        unimplemented!("Just a SQL generation test")
+    }
+
+    fn moving_state_inverse(
+        _current: Self::State,
+        _arg: Self::Args,
+        _fcinfo: pg_sys::FunctionCallInfo,
+    ) -> Self::MovingState {
+        unimplemented!("Just a SQL generation test")
+    }
+
+    fn combine(
+        _first: Self::State,
+        _second: Self::State,
+        _fcinfo: pg_sys::FunctionCallInfo,
+    ) -> Self::State {
+        unimplemented!("Just a SQL generation test")
     }
 }
